requirements.txt
beautifulsoup4
requests
pandas
streamlit
pillow
matplotlib
selenium
